requirement.txt
langchain==0.2.16
langchain-core==0.2.38
langchain-community==0.2.16
langchain-groq==0.1.9
langchain-cohere==0.2.4
faiss-cpu
streamlit
python-dotenv
cohere